chore: extract styles
