+++ src/elements/simple-icon/src/styles.ts
@@ -1,0 +1,25 @@
+import { pxToRem } from '@uswitch/trustyle.styles'
+import { SxStyleProp } from 'theme-ui'
+
+import { Direction } from '.'
+
+export const iconSize = (size: number | undefined): SxStyleProp =>
+  size
+    ? {
+        height: pxToRem(size),
+        width: pxToRem(size)
+      }
+    : {}
+
+export const rotate = (direction: Direction): SxStyleProp => {
+  switch (direction) {
+    case 'right':
+      return { transform: 'rotate(0.25turn)' }
+    case 'down':
+      return { transform: 'rotate(0.5turn)' }
+    case 'left':
+      return { transform: 'rotate(0.75turn)' }
+    default:
+      return {}
+  }
+}
